+++ src/util/index.js
@@ -1,10 +1,12 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+export * from '#x/BaseDefRef';
 export * from '#x/BaseValueVisitor';
 export * from '#x/IntfDeconstructable';
 export * from '#x/ErrorUtil';
 export * from '#x/StackFrame';
 export * from '#x/StackTrace';
+export * from '#x/VisitDef';
 export * from '#x/VisitRef';
 export * from '#x/VisitResult';